fixed strip checkout route, fixed home page search bar, fixed quiz route, made date components
