+++ client/src/App.css
@@ -26,3 +26,14 @@
   left: 133px;
   top: 673px;
 }
+
+#dateleft {
+  margin-right: 5rem;
+}
+
+#date {
+  display: flex;
+  flex-direction: row;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+}
