+++ client/src/Pages/HomePage.jsx
@@ -1,31 +1,40 @@
 import React from 'react';
-import { Carousel, Container, FormControl, Form, Col } from 'react-bootstrap';
+import {
+  Carousel,
+  Container,
+  FormControl,
+  Form,
+  Col,
+  Button
+} from 'react-bootstrap';
 import SearchBars from '../components/SearchBars';
 import NavigationBar from '../components/NavigationBar';
+import { Link } from 'react-router-dom';
+import Map from '../components/Map/Map';
+import Date from '../components/Date';
 
 const Home = () => {
   return (
     <div>
+      <NavigationBar />
+      {/* <Map /> */}
       <Container className="d-flex flex-column justify-content align-items-center mt-5 bg-white">
         <h1>Wherever the campfire takes you.</h1>
         <h6>
           Discover America’s National Parks to unplug and enjoy a night under
           the stars.
         </h6>
-        <NavigationBar />
         <SearchBars />
-        <p></p>
-        <p></p>
+        <Date />
+        <Button>Find My Trip</Button>
         <div className="bg-secondary d-flex flex-column justify-content align-items-center mt-5 p-3 text-white">
           <h1>Need help finding an adventure?</h1>
           <h6>We can narrow it down to fit your needs.</h6>
-          <button>Take Our Quiz</button>
+          <Button type="submit" as={Link} to="/quiz">
+            Take Our Quiz
+          </Button>
         </div>
-
-        <p></p>
-        <p></p>
         <h1>How It Works</h1>
-
         <svg
           width="2em"
           height="2em"
@@ -43,11 +52,9 @@ const Home = () => {
             d="M6.5 12a5.5 5.5 0 1 0 0-11 5.5 5.5 0 0 0 0 11zM13 6.5a6.5 6.5 0 1 1-13 0 6.5 6.5 0 0 1 13 0z"
           />
         </svg>
-        <p></p>
         <h4>Step 1</h4>
         <p>Choose your getaway</p>
         <p>Pick a park & set the date</p>
-
         <svg
           width="2em"
           height="2em"
@@ -62,11 +69,9 @@ const Home = () => {
             d="M8.202.28a.25.25 0 0 0-.404 0l-.91 1.255a.25.25 0 0 1-.334.067L5.232.79a.25.25 0 0 0-.374.155l-.36 1.508a.25.25 0 0 1-.282.19l-1.532-.245a.25.25 0 0 0-.286.286l.244 1.532a.25.25 0 0 1-.189.282l-1.509.36a.25.25 0 0 0-.154.374l.812 1.322a.25.25 0 0 1-.067.333l-1.256.91a.25.25 0 0 0 0 .405l1.256.91a.25.25 0 0 1 .067.334L.79 10.768a.25.25 0 0 0 .154.374l1.51.36a.25.25 0 0 1 .188.282l-.244 1.532a.25.25 0 0 0 .286.286l1.532-.244a.25.25 0 0 1 .282.189l.36 1.508a.25.25 0 0 0 .374.155l1.322-.812a.25.25 0 0 1 .333.067l.91 1.256a.25.25 0 0 0 .405 0l.91-1.256a.25.25 0 0 1 .334-.067l1.322.812a.25.25 0 0 0 .374-.155l.36-1.508a.25.25 0 0 1 .282-.19l1.532.245a.25.25 0 0 0 .286-.286l-.244-1.532a.25.25 0 0 1 .189-.282l1.508-.36a.25.25 0 0 0 .155-.374l-.812-1.322a.25.25 0 0 1 .067-.333l1.256-.91a.25.25 0 0 0 0-.405l-1.256-.91a.25.25 0 0 1-.067-.334l.812-1.322a.25.25 0 0 0-.155-.374l-1.508-.36a.25.25 0 0 1-.19-.282l.245-1.532a.25.25 0 0 0-.286-.286l-1.532.244a.25.25 0 0 1-.282-.189l-.36-1.508a.25.25 0 0 0-.374-.155l-1.322.812a.25.25 0 0 1-.333-.067L8.203.28zM8 2.5a5.5 5.5 0 1 0 0 11 5.5 5.5 0 0 0 0-11z"
           />
         </svg>
-        <p></p>
         <h4>Step 2</h4>
         <p>Browse fun activities</p>
         <p>See our recommended activities for your campsite</p>
-        <p></p>
         <svg
           width="2em"
           height="2em"
@@ -81,11 +86,9 @@ const Home = () => {
           />
           <path d="M7 13h2v3H7v-3z" />
         </svg>
-        <p></p>
         <h4>Step 3</h4>
         <p>Embrace the outdoors</p>
         <p>Unplug, relax, and enjoy</p>
-
         <div className="d-flex flex-column justify-content align-items-center mt-3 p-3 text-dark">
           <h4>Browse these articles for some inspiration and information:</h4>
         </div>
