+++ client/src/components/Quiz.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Quiz = () => {
+  return (
+    <div>
+      <h1>SIKKEEEEEEE</h1>
+      <button>Take Our Quiz </button>
+    </div>
+  );
+};
+
+export default Quiz;
