--- client/src/components/Quiz/Quiz.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Quiz = () => {
-   render  (
-   <div>
-       <h1>Need help finding an adventure?</h1>
-       <h3>We can narrow it down to fit your needs</h3>
-       <button>Take Our Quiz </button>
-   </div>
-   )
-}
-
-export default Quiz
