+++ client/src/components/StripeCheckout.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import StripeCheckout from 'react-stripe-checkout';
+import { Container } from 'react-bootstrap';
+
+const StripeCheckoutButton = ({ price }) => {
+  const priceForStripe = price * 100;
+  const publishableKey = process.env.STRIPE_PUBLISHABLE_KEY;
+
+  const onToken = (token) => {
+    console.log(token);
+    alert('Payment Succesful!');
+  };
+
+  return (
+    <StripeCheckout
+      label="Pay Now"
+      name="Campo "
+      billingAddress
+      image=""
+      description={`Your total is $${price}`}
+      amount={priceForStripe}
+      panelLabel="Pay Now"
+      token={onToken}
+      stripeKey={publishableKey}
+    />
+  );
+};
+
+export default StripeCheckoutButton;
